+++ AutoShop.DAL/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -50,11 +50,11 @@ namespace AutoShop.DAL.Migrations
 
             modelBuilder.Entity("AutoShop.Domain.Entity.Car", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<byte[]>("Avatar")
                         .HasColumnType("bytea");
